feat: Add functionality to hide Storyteller item
This commit adds the ability to hide a Storyteller item in the HomeScreen and TabScreen. When a user interacts with a specific Storyteller item, it will now be hidden from the UI. This logic change is implemented in the HomeViewModel and TabViewModel classes.

The code changes include:
- Added a new function `hideStorytellerItem` in both HomeViewModel and TabViewModel.
- Updated the `HomeScreen` and `TabScreen` functions to call `viewModel.hideStorytellerItem` when interacting with a Storyteller item.
- Modified the `PageItemStorytellerDelegate` class to call the provided callback function `onShouldHide` when there is an error or no data for a Storyteller item.

These changes improve user experience by allowing them to hide empty or errored Storyteller items from their view.