feat: Improve accessibility by adding content descriptions to icons in MainScreen
This commit adds content descriptions to the search and account icons in the MainScreen component. The search icon now has a content description of "Open Search" and the account icon has a content description of "Open Account". This improves accessibility for users who rely on screen readers or other assistive technologies.