feat: Add top border to BottomNavigationBar
This commit adds a new feature to the BottomNavigationBar component. It introduces a top border using the `borderTop` extension function from `ComposeExtensions.kt`. The border is drawn with a width of 0.5dp and uses the color defined in `LocalStorytellerColorsPalette.current.border`. This change enhances the visual appearance of the navigation bar.