feat: Add reloadCurrentTabSignal to TabLayout
This commit adds a new variable `reloadCurrentTabSignal` to the TabLayout in order to track the index of the currently selected tab that needs to be reloaded. This variable is used in conjunction with the `NavigationInterceptor.None` condition to trigger a reload of the current tab when it is selected again.